test bench create

+++ central_control/Core/Inc/motor_control.h
@@ -8,9 +8,11 @@
 #ifndef INC_MOTOR_CONTROL_H_
 #define INC_MOTOR_CONTROL_H_
 
+#include "main.h"
+
 extern TIM_HandleTypeDef htim4;
 
-void motor_control (uint16_t mode);
+void motor_control (uint8_t mode);
